resizer-n-uploader/requirements.txt
gunicorn
flask
Pillow
requests
cmake
dlib
face_recognition
cloudinary
